Attempt to fix permissions for trivy scan on main (#867)
While the scan works on PRs, we have permssions related
failures for the run on the main branch, this attempts to fix that.

Similarly, this moves to always scan the image on main, as a way
to double check there are no new blocking CVEs more regularly than
just when the image is bumped.